Fixed a couple more bugs, more to take care of though

+++ src/handleSignal.c
@@ -1,34 +1,45 @@
 #include "../include/handleSignal.h"
 
+static bool gonnaDieSoon = false;
+
+void alarmHandler(int signo) {
+    sendSignal(getpid(), "GP SIGCONT"); 
+    killpg(thisOpt->child_pgid, SIGCONT);
+    gonnaDieSoon = false;
+    raise(SIGCONT);
+}
 
 /**
  * @brief It handles the father SIGINT interrupt
  * 
  */
 void handlerFather(int signo){
-    char key[1]; 
-    //STOP children
-    writeInLog(RECV_SIGNAL, "SIGINT");
-    killpg(thisOpt->child_pgid, SIGSTOP);
-    
-    write(STDOUT_FILENO, "|| [Y/y] to proceed | any other key to continue ||", 50); 
-    read(STDIN_FILENO, &key, 1);
 
-    if (strcmp("Y", key) == 0 || strcmp("y", key) == 0){
+    if (!gonnaDieSoon) {
+        gonnaDieSoon = true;
+
+        //STOP children
+        writeInLog(SEND_SIGNAL, "SIGSTOP");
+        killpg(thisOpt->child_pgid, SIGSTOP);
+
+        fflush(stdout);
+        write(STDOUT_FILENO, "____ Ctrl+C again to terminate, the program will continue in 2 sec ____", 71);
+
+        alarm(2);
+        sleep(2);
+    }
+    else {
+        alarm(0);
         sendSignal(getpid(), "GP SIGINT"); 
         writeInLog(RECV_SIGNAL, "SIGINT"); 
         killpg(thisOpt->child_pgid, SIGINT);
         exit(1);
     }
-    
-    //CONTINUE children
-    sendSignal(getpid(), "GP SIGCONT"); 
-    killpg(thisOpt->child_pgid, SIGCONT);
 
 }
 
 void handlerChild_sigCont(int signo){
-    writeInLog(RECV_SIGNAL, "SIGCONT"); 
+    writeInLog(RECV_SIGNAL, "SIGCONT");
 }
 
 void handlerChild_sigInt(int signo){
@@ -48,17 +59,23 @@ int setSignal(Options *opt){
     if (opt->original_process) {
         act.sa_handler = handlerFather;
         if (sigaction(SIGINT, &act, NULL) < 0){
-            perror("Error set father signal"); 
+            perror("Error set father interrupt signal"); 
             return 1; 
-        } 
+        }
+
+        act.sa_handler = alarmHandler;
+        if (sigaction(SIGALRM, &act, NULL) < 0){
+            perror("Error set father alarm signal"); 
+            return 1; 
+        }
     }
     else{
-        act.sa_handler = handlerChild_sigCont; 
-        if (sigaction(SIGCONT, &act, NULL) < 0){
-            perror("Error set child signal, SIGCONT"); 
-            return 1; 
+        // act.sa_handler = handlerChild_sigCont; 
+        // if (sigaction(SIGCONT, &act, NULL) < 0){
+        //     perror("Error set child signal, SIGCONT"); 
+        //     return 1; 
 
-        }
+        // }
 
         act.sa_handler= handlerChild_sigInt; 
         if (sigaction(SIGINT, &act, NULL) < 0){
